First commit for the NEST Thermostat Simulation and the IoT Transformation Repository Interface
The nest simulator project simulates the NEST Thermostat. The aim of this simulation is to demonstrate how the hub will achieve interoperability when receiving messages from the different IoT devices.

The repository interface project takes the device ID from the hub and uses it to query the MySQL database which contains the mapping between IoT devices and the transformation class name.